+++ iceoryx_hoofs/test/moduletests/test_shared_memory.cpp
@@ -1,5 +1,5 @@
 // Copyright (c) 2019 by Robert Bosch GmbH. All rights reserved.
-// Copyright (c) 2021 by Apex.AI Inc. All rights reserved.
+// Copyright (c) 2021 - 2022 by Apex.AI Inc. All rights reserved.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -58,6 +58,7 @@ class SharedMemory_Test : public Test
             .name(name)
             .accessMode(iox::posix::AccessMode::READ_WRITE)
             .openMode(openMode)
+            .filePermissions(cxx::perms::owner_all)
             .size(128)
             .create();
     }